refactor(main): rename quick-clips.ts to quick-clips-config.ts
The module has held only exportQuickClipsConfig and importQuickClipsConfig
since step 1 moved scanning to src/shared and the launcher's tool opening
went. The name now says what is in it. The test drops a leftover electron
mock that nothing asserted on.
