+++ src/main/clipboard/quick-clips-config.ts
